Moving State and DOM visual elements to own files
